feat: 리뷰 목록 조회 API 구현

+++ api/store/src/main/java/com/backtothefuture/store/dto/request/ReviewCreateRequest.java
@@ -9,7 +9,7 @@ public record ReviewCreateRequest(
         Long storeId,
 
         @NotNull
-        int ratingCount, // TODO 더블로 수정
+        Double ratingCount,
 
         String content,
         MultipartFile imageFile
